@@ -6,16 +6,14 @@ use ricwein\Tokenizer\InputSymbols\Delimiter;
 
 abstract class BaseToken
 {
-    protected int $line;
-    protected ?Delimiter $delimiter;
+    public function __construct(protected ?Delimiter $delimiter, private readonly int $line = 1)
+    {
+    }
 
     abstract public function __toString(): string;
 
     abstract public function content(): string;
 
-    /**
-     * @return Delimiter|null
-     */
     public function delimiter(): ?Delimiter
     {
         return $this->delimiter;
@@ -26,20 +24,12 @@ abstract class BaseToken
         return $this->line;
     }
 
-    /**
-     * @param Delimiter|null $delimiter
-     * @return $this
-     */
     public function setDelimiter(?Delimiter $delimiter): self
     {
         $this->delimiter = $delimiter;
         return $this;
     }
 
-    /**
-     * @param string|null $delimiter
-     * @return bool
-     */
     public function isDelimiter(?string $delimiter): bool
     {
         if ($this->delimiter === null) {
@@ -49,9 +39,6 @@ abstract class BaseToken
         return $this->delimiter->is($delimiter);
     }
 
-    /**
-     * @return bool
-     */
     public function isContextSwitching(): bool
     {
         if ($this->delimiter === null) {
