@@ -12,30 +12,23 @@ use UnexpectedValueException;
 
 class Tokenizer
 {
-    protected const MAX_DEPTH = 128;
-
     /** @var Delimiter[] */
     private array $delimiterToken = [];
 
     /** @var Block[] */
     private array $blockToken = [];
 
-    // config flags
-    private int $maxDepth = self::MAX_DEPTH;
-    private bool $disableAutoTrim = false;
-
     /**
-     * Tokenizer constructor.
      * @param Delimiter[] $delimiterToken
      * @param Block[] $blockToken
-     * @param array $config
      * @throws UnexpectedValueException
      */
-    public function __construct(array $delimiterToken, array $blockToken, array $config = [])
+    public function __construct(
+        array                   $delimiterToken,
+        array                   $blockToken,
+        private readonly Config $config = new Config()
+    )
     {
-        // read type-safe config entries
-        $this->parseConfig($config);
-
         $takenSymbols = [];
 
         foreach ($delimiterToken as $delimiter) {
@@ -46,19 +39,21 @@ class Tokenizer
             $this->delimiterToken[] = $delimiter;
         }
 
-        // sort delimiters by longest first, this assures,
-        // that multichar delimiters can match before a similar singlechar delimiter
-        // is check which could also match
-        usort($this->delimiterToken, function (Delimiter $lhs, Delimiter $rhs): int {
-            return $rhs->length() - $lhs->length();
-        });
+        /**
+         * Sort delimiters by longest first, this assures that multichar delimiters can match
+         * before a similar single-char delimiter is checked (which could also match)
+         */
+        usort(
+            $this->delimiterToken,
+            static fn(Delimiter $lhs, Delimiter $rhs): int => $rhs->length() - $lhs->length()
+        );
 
         foreach ($blockToken as $block) {
             if (in_array($block->open()->symbol(), $takenSymbols, true)) {
-                throw new UnexpectedValueException("Found duplicated symbol in Tokenizer for '{$block->symbolOpen()}' Block open-symbol.", 500);
+                throw new UnexpectedValueException("Found duplicated symbol in Tokenizer for '{$block->open()->symbol()}' Block open-symbol.", 500);
             }
             if (in_array($block->close()->symbol(), $takenSymbols, true)) {
-                throw new UnexpectedValueException("Found duplicated symbol in Tokenizer for '{$block->symbolClose()}' Block close-symbol.", 500);
+                throw new UnexpectedValueException("Found duplicated symbol in Tokenizer for '{$block->close()->symbol()}' Block close-symbol.", 500);
             }
 
             $takenSymbols[] = $block->open()->symbol();
@@ -68,26 +63,12 @@ class Tokenizer
             $this->blockToken[] = $block;
         }
 
-        usort($this->blockToken, function (Block $lhs, Block $rhs): int {
-            return $rhs->open()->length() - $lhs->open()->length();
-        });
+        usort(
+            $this->blockToken,
+            static fn(Block $lhs, Block $rhs): int => $rhs->open()->length() - $lhs->open()->length()
+        );
     }
 
-    private function parseConfig(array $config): void
-    {
-        if (array_key_exists('maxDepth', $config) && is_int($config['maxDepth'])) {
-            $this->maxDepth = $config['maxDepth'];
-        }
-        if (array_key_exists('disableAutoTrim', $config) && is_bool($config['disableAutoTrim'])) {
-            $this->disableAutoTrim = $config['disableAutoTrim'];
-        }
-    }
-
-    /**
-     * @param string $input
-     * @param int $startLine
-     * @return TokenStream
-     */
     public function tokenize(string $input, int $startLine = 1): TokenStream
     {
         return new TokenStream($this->process($input, 0, $startLine));
@@ -95,16 +76,13 @@ class Tokenizer
 
     /**
      * splits given string into blocks and symbols
-     * @param string $input
-     * @param int $depth
-     * @param int $line
      * @return BlockToken[]|Token[]
      */
     private function process(string $input, int $depth, int $line): array
     {
         // abort tokenizing after reaching the max block depth
         // just return the input string as the remaining symbol
-        if ($this->maxDepth > 0 && $depth >= $this->maxDepth) {
+        if ($this->config->maxDepth > 0 && $depth >= $this->config->maxDepth) {
             return [new Token($input, null, $line)];
         }
 
@@ -120,7 +98,7 @@ class Tokenizer
         /** @var Delimiter|null $lastDelimiter */
         $lastDelimiter = null;
 
-        /** @var string $lastChar single char from previous iteration */
+        /** single char from previous iteration */
         $lastChar = '';
 
         $lastOffset = 0;
@@ -128,11 +106,11 @@ class Tokenizer
 
         foreach (str_split($input) as $offset => $char) {
             if ($lastChar === PHP_EOL) {
-                $line += 1;
+                ++$line;
             }
             $lastChar = $char;
 
-            // fast forward for match multi-char delimiters
+            // fast-forward for match multi-char delimiters
             if ($lastOffset > $offset) {
                 continue;
             }
@@ -195,7 +173,7 @@ class Tokenizer
                     if ($lastOffset < $offset) {
 
                         $prefix = substr($input, $lastOffset, $offset - $lastOffset);
-                        if (!$this->disableAutoTrim) {
+                        if (!$this->config->disableAutoTrim) {
                             $prefix = trim($prefix);
                         }
                         if (!empty($prefix)) {
@@ -218,7 +196,7 @@ class Tokenizer
             }
 
             // scanning for delimiters inside blocks is not required, since it's done inside
-            // sub- process() calls, where the currently open block is the main block
+            // sub-process() calls, where the currently open block is the main block
             if (!empty($openBlocks)) {
                 continue;
             }
@@ -231,11 +209,11 @@ class Tokenizer
                         // only keep symbol, if it's not already processed before
                         // e.g. if the previous symbol was a block
                         $content = substr($input, $lastOffset, $offset - $lastOffset);
-                        if (!$this->disableAutoTrim) {
+                        if (!$this->config->disableAutoTrim) {
                             $content = rtrim($content);
                         }
 
-                        // encounter of symbol directly after an block (no delimiter in between)
+                        // encounter of symbol directly after a block (no delimiter in between)
                         if ($lastSymbol instanceof BlockToken) {
 
                             if (!empty($content)) {
@@ -249,7 +227,7 @@ class Tokenizer
                             }
 
                             // we need to reset the last-symbol, since we processed the
-                            // current symbol as an block-suffix
+                            // current symbol as a block-suffix
                             $lastSymbol = null;
                         } else {
                             $lastSymbol = new Token($content, $lastDelimiter, $line);
@@ -258,7 +236,7 @@ class Tokenizer
 
                     } else {
 
-                        // set last-symbol since we encountered an delimiter, but
+                        // set last-symbol since we encountered a delimiter, but
                         // because we can skip the symbol, we set last-symbol to null
                         $lastSymbol = null;
 
@@ -274,10 +252,10 @@ class Tokenizer
         }
 
         // handle remaining tokens
-        if (!$this->disableAutoTrim) {
+        if (!$this->config->disableAutoTrim) {
             $remaining = ltrim($remaining, ' ');
         }
-        if (strlen($remaining) > 0) {
+        if ($remaining !== '') {
             if ($lastSymbol instanceof BlockToken) {
 
                 if ($lastSymbol->block()->splitAffixIntoSymbols()) {
